requirements.txt
flask
pandas
matplotlib
seaborn
google-generativeai
openpyxl
